Seperated ther Serilization Data from the task serializer
