+++ Assets/Virtual Agents Framework/Runtime/Scripts/BehaviorTree/Abstract Behaviour Tree/NodeImplementations/DebugTask.cs
@@ -16,11 +16,11 @@ namespace i5.VirtualAgents.TaskSystem.AgentTasks
             FinishTask();
         }
 
-        public void Serialize(TaskSerializer serializer)
+        public void Serialize(SerializationDataContainer serializer)
         {
         }
 
-        public void Deserialize(TaskSerializer serializer)
+        public void Deserialize(SerializationDataContainer serializer)
         {
         }
     }
