+++ Assets/Virtual Agents Framework/Runtime/Scripts/BehaviorTree/Visual Behaviour Tree/SerializationDataContainer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7be94d51cfff3ce4bb2405e3ce87546b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
